Fixes #3177

Improve performance while sharing (adding and removing a user from) an album.

In this PR:
- Don't return assets on the update album endpoint response
- Don't return assets on the add album user endpoint response
- Don't load assets from the database when removing a user (mobile app needs changes before assets can be skipping when adding a new user)
- Don't ask for assets when refreshing the album info (web)

Before vs After Tests (on an album with 10k assets)

1.9 seconds to add a user (before) vs 800ms to add a user (after)
1.7 seconds to remove a user (before) vs 600ms to remove a user (after)

![image](https://github.com/immich-app/immich/assets/4334196/e522ee80-5389-4a95-9d07-892eed7b85c9)

Future Improvements:
- Adding/removing assets to an album should be done without loading all the assets into memory
- Loading all the assets into memory is currently required to get the album asset count. This can be avoided, but may require using query builder or running two selects.

Tested Scenarios:
- Add a user to an album
- Remove a user from an album
- Leaving a shared album
- Creating a new album