@jrasm91 2023-08-15
> Tested on mobile and it works fine.
> 
> Ah wait, I have to run this PR on mobile as well? right

There are no mobile changes in this code, so as long as the mobile app (on main/release) can create an album, add a user, remove a user, update the title/description, etc. we should be good.